Rename tests to testing
